requirements.txt
Django==4.2.7
Pillow==10.1.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0

# Django REST Framework and JWT
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3

# Database
psycopg2-binary==2.9.7

# Background tasks
celery==5.3.4
redis==5.0.1

# Additional utilities
django-extensions==3.2.3
